client tests working, added common package
